gpt_text2video/backend/requirements.txt
flask
openai
gTTS
moviepy
Pillow
requests
